# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test MSIB IT Backend - Edo Arya Hermawan",
        "description": "Membuat aplikasi dengan konsep REST-API pada sisi backend untuk memenuhi kebutuhan sistem perhitungan lembur (overtime)",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/overtime-pays/calculate": {
            "get": {
                "tags": [
                    "Overtime Pay"
                ],
                "summary": "Calculate overtime pay",
                "operationId": "9ab4e6a1ff633a3855424efc1aa94f6f",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "YYYY-MM"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Overtime pay calculated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}